+ Creado el patron para Timothy - Fallado al intentar mandar el patron por parametro a una función, da errores porque creo que no es posible mandar parametros que no dependan de T

+++ Muphic/Mu/source/include/Figuras.h
@@ -42,7 +42,7 @@ class Figuras
     protected:
 		int sheetWidth;
 		int sheetHeight;
-		int vistosidadTotal;
+		float vistosidadTotal;
 
     private:
 		list<Figura*> figuras;
